+++ Eco_base/STM32F405VG/Project/MDK-ARM/core/quick_math.h
@@ -3,7 +3,7 @@
 
 #include "stm32f4xx.h"
 
-float get_median_of_five(float a, float b, float c, float d, float e);
+s16 get_median_of_five(s16 a, s16 b, s16 c, s16 d, s16 e);
 
 extern __inline int16_t s16_cap(int16_t in_num, int16_t upper_bound, int16_t lower_bound);
 extern __inline uint8_t u16_sqrt(uint16_t in_num);
